+++ apps/oxlint/src-js/plugins/load.ts
@@ -1,14 +1,14 @@
 import { pathToFileURL } from 'node:url';
 
 import { createContext } from './context.js';
-import { getErrorMessage } from './utils.js';
+import { getErrorMessage } from '../utils/utils.js';
 
 import type { Writable } from 'type-fest';
 import type { Context } from './context.ts';
 import type { JsonValue } from './json.ts';
 import type { RuleMeta } from './rule_meta.ts';
 import type { AfterHook, BeforeHook, Visitor, VisitorWithHooks } from './types.ts';
-import type { SetNullable } from './utils.ts';
+import type { SetNullable } from '../utils/types.ts';
 
 const ObjectKeys = Object.keys;
