--- apps/oxlint/src-js/plugins/utils.ts
@@ -1,75 +1,0 @@
-/**
- * Get error message from an error.
- *
- * `err` is expected to be an `Error` object, but can be anything.
- *
- * * If it's an `Error`, the error message and stack trace is returned.
- * * If it's another object with a string `message` property, `message` is returned.
- * * Otherwise, a generic "Unknown error" message is returned.
- *
- * This function will never throw, and always returns a non-empty string, even if:
- *
- * * `err` is `null` or `undefined`.
- * * `err` is an object with a getter for `message` property which throws.
- * * `err` has a getter for `stack` or `message` property which returns a different value each time it's accessed.
- *
- * @param err - Error
- * @returns Error message
- */
-export function getErrorMessage(err: unknown): string {
-  try {
-    if (err instanceof Error) {
-      // Note: `stack` includes the error message
-      const { stack } = err;
-      if (typeof stack === 'string' && stack !== '') return stack;
-    }
-
-    const { message } = err as { message?: unknown };
-    if (typeof message === 'string' && message !== '') return message;
-  } catch {}
-
-  return 'Unknown error';
-}
-
-/**
- * Assert a value is of a certain type.
- *
- * Has no runtime effect - only for guiding the type-checker.
- * Minification removes this function and all calls to it, so it has zero runtime cost.
- *
- * @param value - Value
- */
-// oxlint-disable-next-line no-unused-vars
-export function assertIs<T>(value: unknown): asserts value is T {}
-
-/**
- * Assert a value is not `null` or `undefined`.
- *
- * In release builds, is a no-op. Only does runtime checks in debug builds.
- * Minification removes this function and all calls to it in release builds, so it has zero runtime cost.
- *
- * @param value - Value
- */
-// oxlint-disable-next-line no-unused-vars
-export function assertIsNonNull<T>(value: T | null | undefined): asserts value is T {
-  if (!DEBUG) return;
-
-  if (value === null || value === undefined) {
-    // oxlint-disable-next-line typescript/restrict-template-expressions
-    throw new Error(`Expected non-null value, got ${value}`);
-  }
-}
-
-/**
- * Utility type to make specified properties of a type nullable.
- *
- * @example
- * ```ts
- * type Foo = { str: string, num: number };
- * type FooWithNullableNum = SetNullable<Foo, 'num'>;
- * // { str: string, num: number | null }
- * ```
- */
-export type SetNullable<BaseType, Keys extends keyof BaseType = keyof BaseType> = {
-  [Key in keyof BaseType]: Key extends Keys ? BaseType[Key] | null : BaseType[Key];
-};
